Connections the dots, defining the application flow
During the integration tests lots of points were found where wrong
decisions leaded to code refactoring and re-tests

Luckly as the TDD was applied most the logic became intact and only
small parts of the code had to be changed to accommodate the fixings

> Redefine the return types to keep state

The state was never a concern in the begining of project keeping in mind
would be straightforward add the state concept.

The big mistake here was not thing about pass the state back and forth

After realize that mistake all tests were fixed, was not a huge mental
work but physical one reduced by the pattern matching which make the
life really easier

> Adhere to Module.main/1 specification

The main module expects a [] instead a %{}, even forcing a map a default
list is provided when main is called without parameters.

Testing stdio from external system like SO is hard, that's why I abort
from now. I want back to this soon.